Add viewmodel
