--- ImageEditor/src/main/kotlin/org/wordpress/android/imageeditor/HomeFragment.kt
@@ -1,16 +1,0 @@
-package org.wordpress.android.imageeditor
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-import androidx.fragment.app.Fragment
-
-class HomeFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.home_fragment, container, false)
-}
