+++ ImageEditor/src/main/kotlin/org/wordpress/android/imageeditor/home/HomeFragment.kt
@@ -1,0 +1,25 @@
+package org.wordpress.android.imageeditor.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProviders
+import org.wordpress.android.imageeditor.R
+
+class HomeFragment : Fragment() {
+    private lateinit var homeViewModel: HomeViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? = inflater.inflate(R.layout.home_fragment, container, false)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        homeViewModel = ViewModelProviders.of(this).get(HomeViewModel::class.java)
+    }
+}
