+++ ImageEditor/src/main/kotlin/org/wordpress/android/imageeditor/home/HomeViewModel.kt
@@ -1,0 +1,5 @@
+package org.wordpress.android.imageeditor.home
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel()
