+++ build.gradle
@@ -119,6 +119,7 @@ ext {
     appCompatVersion = '1.0.2'
     constraintLayoutVersion = '1.1.3'
     uCropVersion = '2.2.2'
+    lifecycleExtensionsVersion = '2.0.0'
 
     // testing
     jUnitVersion = '4.12'
